+++ spec/wavefront-cli/webhook_spec.rb
@@ -24,4 +24,5 @@ describe "#{word} command" do
                                   matchingMethod: 'CONTAINS' }],
                         sort: { field: 'id', ascending: true } },
               headers: JSON_POST_HEADERS)
+  test_list_output(word)
 end