+++ HISTORY.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.0.0 (06/03/2019)
+* Drop support for Ruby 2.2. This is the only breaking change.
+* New, improved `human` output. Finally fixes bug where heavily
+  indented columns could run into one another.
+* Move `id` and `name` to the top of objects in long listings.
+* Improved output testing
+
 ## 2.18.0 (22/02/2019)
 * Add `usergroup` command, and extend `user` command to cover new
   RBAC features.